rename and reorder Payments tab
