+++ app/components/views/LNPage/SendTab/DecodedPayRequest/EmptyDescription/EmptyDescription.jsx
