+++ app/components/views/LNPage/SendTab/DecodedPayRequest/EmptyDescription/index.js
@@ -1,0 +1,1 @@
+export { default } from "./EmptyDescription";
