+++ app/components/views/LNPage/SendTab/DecodedPayRequest/ExpiryTime/index.js
@@ -1,0 +1,1 @@
+export { default } from "./ExpiryTime";
