+++ app/components/views/LNPage/SendTab/FailedPayment/index.js
@@ -1,0 +1,1 @@
+export { default } from "./FailedPayment";
