+++ app/components/views/LNPage/SendTab/OutstandingPayment/index.js
@@ -1,0 +1,1 @@
+export { default } from "./OutstandingPayment";
