+++ app/components/views/LNPage/SendTab/Payment/index.js
@@ -1,0 +1,1 @@
+export { default } from "./Payment";
