+++ app/components/views/LNPage/SendTab/index.js
@@ -1,0 +1,1 @@
+export { default as SendTab, SendTabHeader } from "./SendTab";
